Refactored config classes and merged export for .plan and .mac files
