+++ brachyutils/planning/optimization/optim_gurobi.py
@@ -18,8 +18,9 @@ from opentps.core.processing.imageProcessing.sitkImageProcessing import image3DT
 from brachyutils.types import BrachyPlan
 from brachyutils.planning.optimization.optim_utils import (
     BrachyDwellTimeOptim, BrachyDwellTime, get_optimization_roi_bounds, resample_crop_the_mask_or_contour_to_optimGrid,
-    compute_dose_rate_matrices, Optimization_Config
+    compute_dose_rate_matrices
 )
+from brachyutils.planning.optimization.optim_configs import Optimization_Config
 import multiprocessing as mp
 from functools import partial
